Jackpot size and net sales in different regions in Ontario, Canada

This data set contains information about the jackpot size (cad dollar) and net sales and ticket sales. Sales are separated by zip code (first 3 digits) and lotto game. I changed the column name so that they are more informative in lower case. As well, I created new year, month, and day columns via the lubridate package. The original data was provided via an Access to Information Act Request. It is now hosted on Open Science Framework.

jackpot_size

Format

A .csv data frame with the following variables:

zip_code

Originally called FSA (forward sortation area). FSA's are geographical regions defined by the first three digits of postal code (e.g. M1C)

start_date

The start of the sales date, year-month-day format

end_date

The end of the sales date, year-month-day format

game

The specific lotto game. One of Lottario, Lotto 649, Lotto 669.

ticket_sales

Number of ticket sold

net_sales

The total cad dollar amount of sales generated

jackpot_size

How large the jackpot size is in cad dollar

year

Year

month

Month

day

Day

Source

https://osf.io/6qcde/download, downloaded 2020-12-12

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# What is the yearly spending for the Lottario in zone M1B in 2012?

jackpot_size %>%
  filter(year == 2012 & game == "Lottario" & zip_code == "M1B")

#> # A tibble: 366 x 10
#>    zip_code start_date end_date   game  ticket_sales net_sales jackpot_size
#>    <chr>    <date>     <date>     <chr>        <dbl>     <dbl>        <dbl>
#>  1 M1B      2012-01-01 2012-01-07 Lott…           75       233       730000
#>  2 M1B      2012-01-02 2012-01-07 Lott…          167       516       730000
#>  3 M1B      2012-01-03 2012-01-07 Lott…          168       466       730000
#>  4 M1B      2012-01-04 2012-01-07 Lott…          274       841       730000
#>  5 M1B      2012-01-05 2012-01-07 Lott…          195       558       730000
#>  6 M1B      2012-01-06 2012-01-07 Lott…          451      1469       730000
#>  7 M1B      2012-01-07 2012-01-07 Lott…          443      1364       730000
#>  8 M1B      2012-01-08 2012-01-14 Lott…          111       288       840000
#>  9 M1B      2012-01-09 2012-01-14 Lott…          217       535       840000
#> 10 M1B      2012-01-10 2012-01-14 Lott…          210       559       840000
#> # … with 356 more rows, and 3 more variables: year <dbl>, month <dbl>,
#> #   day <int>